requirements.txt
torch==2.0.1
pytorch-lightning==2.0.4
torchmetrics==0.11.4
omegaconf==2.3.0
kornia==0.6.12
wandb==0.15.2
segmentation_models_pytorch==0.3.3
hydra-core==1.3.2
ipython
rasterio
geopandas
ipykernel
matplotlib==3.7.1
scikit-image
scikit-learn
georeader-spaceml
transformers==4.30.2
datasets==2.13.1
numpy==1.25.0
spectral
pysolar
mock